Update TwitterProfanityProblem
An updated version of the previous program that includes additional processing to catch some instances of racial slurs that may be misspelled or written with special characters
we first define a regular expression for each racial slur in the set to catch instances of the slur that may be misspelled or written with special characters. We then use the findall() function to count the number of matches for each regular expression in the tweet. We add this count to the total number of racial slurs identified by the Spacy NLP model to get a more complete count of racial slurs in the tweet.